+++ src/main/java/me/flotsam/frettler/engine/ScaleInterval.java
@@ -21,32 +21,30 @@ import lombok.Getter;
 
 public enum ScaleInterval {
   //@formatter:off
-  P1("P1", 0, true, false),
-  m2("m2", 1, false, true),
-  M2("M2", 2, false, false),
-  m3("m3", 3, false, true),
-  M3("M3", 4, false, false),
-  P4("P4", 5, true, false),
+  P1("P1", 0, true),
+  m2("m2", 1, false),
+  M2("M2", 2, false),
+  m3("m3", 3, false),
+  M3("M3", 4, false),
+  P4("P4", 5, true),
 
-  d5("d5", 6, false, true),
+  d5("d5", 6, false),
 
-  P5("P5", 7, true, false),
-  m6("m6", 8, false, true),
-  M6("M6", 9, false, false),
-  m7("m7", 10, false, true),
-  M7("M7", 11, false, false),
-  P8("P8", 12, true, false),
+  P5("P5", 7, true),
+  m6("m6", 8, false),
+  M6("M6", 9, false),
+  m7("m7", 10, false),
+  M7("M7", 11, false),
+  P8("P8", 12, true),
   
-  m9("m9", 13, false, true),
-  M9("M9", 14, false, false),
-  m10("m10", 15, false, false),
-  M10("M10", 16, false, false),
-  m11("M11", 18, false, true),
-  M11("M11", 18, false, false);
+  m9("m9", 13, false),
+  M9("M9", 14, false),
+  m10("m10", 15, false),
+  M10("M10", 16, false),
+  m11("M11", 18, false),
+  M11("M11", 18, false);
   //@formatter:on
 
-  @Getter
-  private boolean flat;
   @Getter
   private boolean perfect;
   @Getter
@@ -54,10 +52,9 @@ public enum ScaleInterval {
   @Getter
   private String label;
 
-  private ScaleInterval(String label, int semiTones, boolean perfect, boolean flat) {
+  private ScaleInterval(String label, int semiTones, boolean perfect) {
     this.label = label;
     this.semiTones = semiTones;
     this.perfect = perfect;
-    this.flat = flat;
   }
 }